backend/requirements.txt
Flask==3.0.2
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.1
pymongo==4.6.1
dnspython==2.6.1 